build: upgrade ESLint v8+typescript-eslint v6 to v9/v8
- Upgrade eslint 8.54.0 (EOL exact pin) to ^9.36.0
- Upgrade @typescript-eslint/* from ^6.12.0 to ^8.37.0
- Add @eslint/eslintrc ^3.3.1 for FlatCompat adapter
- Bump eslint-plugin-react 7.33.2 to ^7.37.0 (ESLint v9 peer support)
- Migrate .eslintrc.js to eslint.config.mjs (flat config)
- Remove .eslintignore (patterns in globalIgnores)
- Resolves GHSA-3ppc-4f35-3m26 GHSA-23c5-xmqv-rm74 GHSA-7r86-cg39-jmmj

Closes #96
