+++ eslint.config.mjs
@@ -1,0 +1,98 @@
+// @ts-check
+import { dirname } from "path"
+import { fileURLToPath } from "url"
+import { FlatCompat } from "@eslint/eslintrc"
+import { defineConfig, globalIgnores } from "eslint/config"
+import tsPlugin from "@typescript-eslint/eslint-plugin"
+import tsParser from "@typescript-eslint/parser"
+import importPlugin from "eslint-plugin-import"
+import tailwindcss from "eslint-plugin-tailwindcss"
+import fs from "fs"
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
+
+const compat = new FlatCompat({ baseDirectory: __dirname })
+
+function getDirectoriesToSort() {
+  const ignoredSortingDirectories = [".git", ".next", ".vscode", "node_modules"]
+  return getDirectories(process.cwd()).filter((f) => !ignoredSortingDirectories.includes(f))
+}
+
+function getDirectories(path) {
+  return fs.readdirSync(path).filter(function (file) {
+    return fs.statSync(path + "/" + file).isDirectory()
+  })
+}
+
+export default defineConfig([
+  globalIgnores([".next/**", "node_modules/**"]),
+
+  // Next.js + Prettier via FlatCompat (no native flat config export)
+  ...compat.extends("next/core-web-vitals", "prettier"),
+
+  // Tailwind via native flat config
+  ...tailwindcss.configs["flat/recommended"],
+
+  // TypeScript-ESLint + import/order rules
+  {
+    files: ["**/*.{ts,tsx,js,jsx,mjs,cjs}"],
+    plugins: {
+      "@typescript-eslint": tsPlugin,
+      import: importPlugin,
+    },
+    languageOptions: {
+      parser: tsParser,
+    },
+    rules: {
+      "tailwindcss/no-custom-classname": "off",
+      "tailwindcss/classnames-order": "off",
+      "testing-library/prefer-screen-queries": "off",
+      "@next/next/no-html-link-for-pages": "off",
+      "@typescript-eslint/no-unused-vars": [
+        "warn",
+        {
+          argsIgnorePattern: "^_",
+          varsIgnorePattern: "^_",
+        },
+      ],
+      "sort-imports": [
+        "warn",
+        {
+          ignoreCase: true,
+          ignoreDeclarationSort: true,
+        },
+      ],
+      "import/order": [
+        1,
+        {
+          groups: ["external", "builtin", "internal", "sibling", "parent", "index"],
+          pathGroups: [
+            ...getDirectoriesToSort().map((singleDir) => ({
+              pattern: `${singleDir}/**`,
+              group: "internal",
+            })),
+            {
+              pattern: "env",
+              group: "internal",
+            },
+            {
+              pattern: "theme",
+              group: "internal",
+            },
+            {
+              pattern: "public/**",
+              group: "internal",
+              position: "after",
+            },
+          ],
+          pathGroupsExcludedImportTypes: ["internal"],
+          alphabetize: {
+            order: "asc",
+            caseInsensitive: true,
+          },
+        },
+      ],
+    },
+  },
+])
